+++ src/game/utils.rs
@@ -2,8 +2,8 @@ use std::sync::{Arc, Mutex};
 
 pub type MessageQueue = Arc<Mutex<Vec<String>>>;
 
-pub fn calculate_bar(value: u64, max_value: u64, precision: u32) -> (String, String) {
-    let percentage: u64 = value * <u32 as Into<u64>>::into(precision) / max_value;
+pub fn calculate_bar(value: u128, max_value: u128, precision: u32) -> (String, String) {
+    let percentage: u128 = value * <u32 as Into<u128>>::into(precision) / max_value;
     let mut filled = String::from("");
     let mut empty = String::from("");
     for i in 0..precision {
